requirements.txt
openai==1.55.2
protobuf==5.28.3
python-dotenv==1.0.1
streamlit==1.40.2
streamlit_carousel==1.1.2
google-generativeai
